feat: tab render error
